+++ src/js/client.ts
@@ -109,8 +109,8 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
   /**
    * Sets up the integrations
    */
-  public setupIntegrations(): void {
-    super.setupIntegrations();
+  protected _setupIntegrations(): void {
+    super._setupIntegrations();
     const tracing = this.getIntegration(ReactNativeTracing);
     const routingName = tracing?.options.routingInstrumentation?.name;
     if (routingName) {